feat: fixes/updates to work with py3.13
