--- fast_mail_parser/__init__.py
@@ -1,8 +1,0 @@
-from .fast_mail_parser import parse_email, ParseError, PyMail, PyAttachment
-
-__all__ = [
-    "parse_email",
-    "ParseError",
-    "PyMail",
-    "PyAttachment",
-]
