+++ src/fast_mail_parser/__init__.py
@@ -1,0 +1,8 @@
+from ._core import ParseError, PyAttachment, PyMail, parse_email
+
+__all__ = [
+    "ParseError",
+    "PyAttachment",
+    "PyMail",
+    "parse_email",
+]
