+++ src/fast_mail_parser/_core.pyi
