+++ taplo.toml
@@ -1,0 +1,20 @@
+[formatting]
+  align_comments        = true
+  align_entries         = true
+  allowed_blank_lines   = 1
+  array_auto_collapse   = false
+  array_auto_expand     = true
+  array_trailing_comma  = true
+  column_width          = 100
+  compact_arrays        = false
+  compact_entries       = false
+  compact_inline_tables = false
+  crlf                  = false
+  indent_entries        = true
+  indent_string         = "  "
+  indent_tables         = true
+  inline_table_expand   = true
+  reorder_arrays        = true
+  reorder_inline_tables = true
+  reorder_keys          = true
+  trailing_newline      = true
